+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `priceDefinition` property to `OrderFormItem` type  on OrderForm.graphql.
+- `priceDefinition` typings to `Checkout.ts`.
 
 ## [2.143.2] - 2021-08-31
 ### Fixed